VirtualContent: initial commit. Implementation+Test
The VirtualContent provides methods to generate random virtual content,
to create an archive this that content and to verify an archive against
the virtual content. It also possible to write the virtual content to the disk
and to print a list of virtual files (item).

VirtualContent uses ByteArrayStream to represents virtual file content (blob).
This simplifies passing virtual file content to the callback method.

A very simple test based on VirtualContent is added.