JNI-Tools initial commit
A lot of hacking around JNI-Tools. The "old" and the "new" version of it.
The old version partially used by some classes. The "new" version not used
yet but tested with a single junit test.

The "new" version contains mapping for classes with
- private methods
- final methods
- static methods
- virtual methods
for interfaces (no void methods supported yet) and non-static and
static fields. All class and field methods are tested.