Stop removing copy of the dynamic libs from the tmp dir
Dynamic libraries (DLL, so) was extracted from jar and copied
to the temp directory for the time jvm runs. This commit leaves
copied libraries in the temporary directory and uses it next time
7-Zip-JBinding must be initialized. This increase 7-Zip-JBinding
avarage initialization time drastically (on windows).

The copied dynamic libraries get found by the build-ref - the random
unique string assigned to the dynamic libararies on each build.